Skjelett for å bruke REST endepunkt i steden for Kafka topic for oppfølgingstatus
